addons/terrain_map_fetcher/python/requirements.txt
rasterio>=1.3
numpy>=1.24
Pillow>=10.0
OpenEXR>=1.3
requests>=2.28

shapely>=2.0
